requirements.txt
ttkbootstrap==1.13.11
pillow==10.4.0
pystray==0.19.5
flask==3.1.1
flask-cors==6.0.1
cryptography==45.0.4